src/storybook/tools/quality/assessment.py
from typing import Dict, Any
from pydantic import Field
from storybook.tools.base import NovelWritingTool

class QualityMetricsTool(NovelWritingTool):
    name: str = Field(default="quality_metrics")
    description: str = Field(default="Track and assess quality metrics")
    
    async def _arun(self, content: Dict[str, Any], metrics: Dict[str, Any]) -> Dict[str, Any]:
        """Assess content quality."""
        return {"quality_metrics": {
            "readability": 0.0,
            "coherence": 0.0,
            "engagement": 0.0,
            "technical": 0.0
        }}

class QualityGateTool(NovelWritingTool):
    name: str = Field(default="quality_gate")
    description: str = Field(default="Manage quality gates")
    
    async def _arun(self, content: Dict[str, Any], requirements: Dict[str, Any]) -> Dict[str, Any]:
        """Check quality gate requirements."""
        return {"gate_assessment": {
            "passed": False,
            "failures": [],
            "recommendations": []
        }}

class QualityVerificationTool(NovelWritingTool):
    name: str = Field(default="quality_verification")
    description: str = Field(default="Verify quality standards")
    
    async def _arun(self, content: Dict[str, Any], standards: Dict[str, Any]) -> Dict[str, Any]:
        """Verify content against quality standards."""
        return {"verification_results": {
            "passed": False,
            "violations": [],
            "suggestions": []
        }}